remove msm8956-v1.1 1958 MHz node
The msm8956 runs with 1.8 GHz in the BIG cluster, if you want to run with more clock you should overvolt it. This clock is cause of random reboots in AOSP roms. Sol, remove it if you want to fix it. Or, if you want to overclock, add 50 ~ 75 mV in the regulator node of the BIG cpu corresponding to this clock. I'm running my suzu with 2.1 GHz stable in big core, but with a cost: +100 mV overvolting. I think the default users should use the 1.8 GHz stock clock for daily use.